tests/wpt/web-platform-tests/dom/abort/addEventListenerAbortController.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" title="Joey Arhar" href="mailto:jarhar@chromium.org">

<!-- This behavior has not been specified yet. See https://github.com/whatwg/dom/issues/911 -->

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>

<script>
test(t => {
  const target = new EventTarget();
  const controller = new AbortController();

  target.addEventListener('testevent', t.step_func(() => {
    assert_unreached('testevent should have been canceled by AbortController');
  }), {signal: controller.signal});

  controller.abort();
  target.dispatchEvent(new Event('testevent'));
}, 'Tests support for EventController to cancel event listeners in addEventListener.');

test(t => {
  const target = new EventTarget();
  const controller = new AbortController();

  target.addEventListener('testevent', t.step_func(() => {
    assert_unreached('testevent should have been canceled by AbortController');
  }), {signal: controller.signal, capture: true});

  controller.abort();
  target.dispatchEvent(new Event('testevent'));
}, 'Tests support for EventController to cancel capture event listeners in addEventListener.');

test(function() {
  let count = 0;
  function handler() {
    count++;
  }
  const et = new EventTarget();
  const controller = new AbortController();
  et.addEventListener('test', handler, { signal: controller.signal });
  et.dispatchEvent(new Event('test'));
  assert_equals(count, 1, "Adding a signal still adds a listener");
  et.dispatchEvent(new Event('test'));
  assert_equals(count, 2, "The listener was not added with the once flag");
  controller.abort();
  et.dispatchEvent(new Event('test'));
  assert_equals(count, 2, "Aborting on the controller removes the listener");
  et.addEventListener('test', handler, { signal: controller.signal });
  et.dispatchEvent(new Event('test'));
  assert_equals(count, 2, "Passing an aborted signal never adds the handler");
}, "Passing an AbortSignal to addEventListener options should allow removing a listener");
</script>
